extract resource field component
